chore(controller): add missing related resource field in job event
