+++ client/starwhale/base/client/models/models.py
@@ -1047,6 +1047,7 @@ class ResponseMessageListRunVo(SwBaseModel):
 class EventVo(SwBaseModel):
     event_type: EventType = Field(..., alias='eventType')
     source: Source
+    related_resource: RelatedResource = Field(..., alias='relatedResource')
     message: str
     data: Optional[str] = None
     timestamp: Optional[int] = None
@@ -1303,10 +1304,10 @@ class ResourcePool(SwBaseModel):
 class EventRequest(SwBaseModel):
     event_type: EventType = Field(..., alias='eventType')
     source: Source
+    related_resource: RelatedResource = Field(..., alias='relatedResource')
     message: str
     data: Optional[str] = None
     timestamp: Optional[int] = None
-    related_resource: RelatedResource = Field(..., alias='relatedResource')
 
 
 class DataConsumptionRequest(SwBaseModel):
