+++ server/controller/src/main/java/ai/starwhale/mlops/domain/event/EventConverter.java
@@ -39,6 +39,10 @@ public class EventConverter {
                 .id(entity.getId())
                 .eventType(entity.getType())
                 .source(entity.getSource())
+                .relatedResource(EventVo.RelatedResource.builder()
+                        .eventResourceType(entity.getResourceType())
+                        .id(entity.getResourceId())
+                        .build())
                 .message(entity.getMessage())
                 .data(entity.getData())
                 .timestamp(entity.getCreatedTime().getTime())
