+++ server/controller/src/main/java/ai/starwhale/mlops/schedule/impl/k8s/reporting/RunEventListener.java
@@ -19,8 +19,8 @@ package ai.starwhale.mlops.schedule.impl.k8s.reporting;
 import ai.starwhale.mlops.api.protocol.event.Event.EventResourceType;
 import ai.starwhale.mlops.api.protocol.event.Event.EventSource;
 import ai.starwhale.mlops.api.protocol.event.Event.EventType;
+import ai.starwhale.mlops.api.protocol.event.Event.RelatedResource;
 import ai.starwhale.mlops.api.protocol.event.EventRequest;
-import ai.starwhale.mlops.api.protocol.event.EventRequest.RelatedResource;
 import ai.starwhale.mlops.domain.event.EventService;
 import ai.starwhale.mlops.schedule.impl.k8s.K8sClient;
 import ai.starwhale.mlops.schedule.impl.k8s.RunExecutorK8s;
